# app/main.py
# generated by fastapi-codegen:
#   filename:  client-master_openapi.json
#   timestamp: 2025-09-17T03:29:58+00:00

from __future__ import annotations

from typing import Optional

from fastapi import FastAPI

from models import (
    HealthCheck,
    JobCancelJobIdPostResponse,
    JobResultJobIdGetResponse,
    JobStatusJobIdGetResponse,
    JobSubmitPostRequest,
    JobSubmitPostResponse,
)

app = FastAPI(
    title="MapReduce Master API",
    version="0.2.0",
)


@app.get(
    "/",
    response_model=HealthCheck,
)
def health_check() -> HealthCheck:
    """
    Health Check
    """
    return HealthCheck(status="ok")


@app.post("/job/cancel/{job_id}", response_model=JobCancelJobIdPostResponse)
def cancel_job(job_id: str) -> JobCancelJobIdPostResponse:
    """
    Cancel a running job
    """
    pass


@app.get("/job/result/{job_id}", response_model=JobResultJobIdGetResponse)
def get_job_result(job_id: str) -> JobResultJobIdGetResponse:
    """
    Get job result
    """
    pass


@app.get("/job/status/{job_id}", response_model=JobStatusJobIdGetResponse)
def get_job_status(job_id: str) -> JobStatusJobIdGetResponse:
    """
    Get job status
    """
    pass


@app.post(
    "/job/submit",
    response_model=None,
    responses={"201": {"model": JobSubmitPostResponse}},
)
def submit_job(body: JobSubmitPostRequest) -> Optional[JobSubmitPostResponse]:
    """
    Submit a new job
    """
    pass