# app/models.py
# generated by fastapi-codegen:
#   filename:  client-master_openapi.json
#   timestamp: 2025-09-17T03:29:58+00:00

from __future__ import annotations

from enum import Enum
from typing import Optional

from pydantic import AnyUrl, BaseModel, Field, confloat


class Error(BaseModel):
    detail: str


class HealthCheck(BaseModel):
    status: Optional[str] = Field(None, example="ok")


class JobSubmitPostRequest(BaseModel):
    data_url: AnyUrl
    code_url: AnyUrl
    job_name: Optional[str] = None


class JobSubmitPostResponse(BaseModel):
    job_id: Optional[str] = None
    status: Optional[str] = Field(None, example="submitted")


class Status(Enum):
    submitted = "submitted"
    running = "running"
    completed = "completed"
    failed = "failed"


class JobStatusJobIdGetResponse(BaseModel):
    job_id: Optional[str] = None
    status: Optional[Status] = None
    progress: Optional[confloat(ge=0.0, le=100.0)] = None


class JobResultJobIdGetResponse(BaseModel):
    job_id: Optional[str] = None
    result_url: Optional[AnyUrl] = None


class JobCancelJobIdPostResponse(BaseModel):
    job_id: Optional[str] = None
    status: Optional[str] = Field(None, example="cancelled")